Add comprehensive NutritionServiceProtocol and implement missing methods
